+++ crates/holochain_cascade/src/authority.rs
@@ -1,6 +1,5 @@
 //! Functions for the various authorities to handle queries
 
-use self::get_agent_activity_query::hashes::GetAgentActivityQuery;
 use self::get_agent_activity_query::must_get_agent_activity::must_get_agent_activity;
 use self::get_entry_ops_query::GetEntryOpsQuery;
 use self::get_links_ops_query::GetLinksOpsQuery;
@@ -8,8 +7,9 @@ use self::{
     get_agent_activity_query::deterministic::DeterministicGetAgentActivityQuery,
     get_record_query::GetRecordOpsQuery,
 };
-
 use super::error::CascadeResult;
+use crate::authority::get_agent_activity_query::hashes::GetAgentActivityHashesQuery;
+use crate::authority::get_agent_activity_query::records::GetAgentActivityRecordsQuery;
 use holo_hash::ActionHash;
 use holo_hash::AgentPubKey;
 use holochain_state::query::link::GetLinksQuery;
@@ -52,26 +52,39 @@ pub async fn handle_get_record(
     Ok(results)
 }
 
-/// Handler for get_agent_activity query to an Activity authority
+/// Handler for get_agent_activity query to an Activity authority.
 #[cfg_attr(feature = "instrument", tracing::instrument(skip(env)))]
 pub async fn handle_get_agent_activity(
     env: DbRead<DbKindDht>,
     agent: AgentPubKey,
     query: ChainQueryFilter,
     options: holochain_p2p::event::GetActivityOptions,
-) -> CascadeResult<AgentActivityResponse<ActionHash>> {
-    let query = GetAgentActivityQuery::new(agent.clone(), query, options);
+) -> CascadeResult<AgentActivityResponse> {
     let results = env
-        .read_async(move |txn| {
+        .read_async(move |txn| -> CascadeResult<AgentActivityResponse> {
             let txn = Txn::from(&txn);
-            let warrants = txn.get_warrants_for_basis(&AnyLinkableHash::from(agent), true)?;
-            let mut r = query.run(txn)?;
+
+            let warrants =
+                txn.get_warrants_for_basis(&AnyLinkableHash::from(agent.clone()), true)?;
+
+            let mut activity_response = if options.include_full_records {
+                // If the caller wanted records, prioritise giving those back.
+                GetAgentActivityRecordsQuery::new(agent, query, options).run(txn)?
+            } else {
+                // Otherwise, just give back the hashes.
+                GetAgentActivityHashesQuery::new(agent, query, options).run(txn)?
+            };
+
             if !warrants.is_empty() {
-                r.warrants = warrants.into_iter().map(|w| w.into_warrant()).collect();
+                // TODO why did we retrieve warrants in the activity query if we're going to overwrite them here?
+                activity_response.warrants =
+                    warrants.into_iter().map(|w| w.into_warrant()).collect();
             }
-            CascadeResult::Ok(r)
+
+            Ok(activity_response)
         })
         .await?;
+
     Ok(results)
 }